+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/environment/StreamExecutionEnvironment.java
@@ -673,18 +673,14 @@ public abstract class StreamExecutionEnvironment {
 	}
 
 	/**
-	 * Creates a new data stream that contains the given elements. The elements must all be of the same type, for
-	 * example, all of the {@link String} or {@link Integer}.
-	 * <p>
-	 * The framework will try and determine the exact type from the elements. In case of generic elements, it may be
-	 * necessary to manually supply the type information via {@link #fromCollection(java.util.Collection,
-	 * org.apache.flink.api.common.typeinfo.TypeInformation)}.
-	 * <p>
+	 * Creates a new data set that contains the given elements. The framework will determine the type according to the 
+	 * based type user supplied. The elements should be the same or be the subclass to the based type. 
+	 * The sequence of elements must not be empty.
 	 * Note that this operation will result in a non-parallel data stream source, i.e. a data stream source with a
 	 * degree of parallelism one.
 	 *
-	 * @param clazz
-	 * 		The base class type in the collection.
+	 * @param type
+	 * 		The based class type in the collection.
 	 * @param data
 	 * 		The array of elements to create the data stream from.
 	 * @param <OUT>
@@ -692,17 +688,17 @@ public abstract class StreamExecutionEnvironment {
 	 * @return The data stream representing the given array of elements
 	 */
 	@SafeVarargs
-	public final <OUT> DataStreamSource<OUT> fromElements(Class<OUT> clazz, OUT... data) {
+	public final <OUT> DataStreamSource<OUT> fromElements(Class<OUT> type, OUT... data) {
 		if (data.length == 0) {
 			throw new IllegalArgumentException("fromElements needs at least one element as argument");
 		}
 
 		TypeInformation<OUT> typeInfo;
 		try {
-			typeInfo = TypeExtractor.getForClass(clazz);
+			typeInfo = TypeExtractor.getForClass(type);
 		}
 		catch (Exception e) {
-			throw new RuntimeException("Could not create TypeInformation for type " + data[0].getClass().getName()
+			throw new RuntimeException("Could not create TypeInformation for type " + type.getName()
 					+ "; please specify the TypeInformation manually via "
 					+ "StreamExecutionEnvironment#fromElements(Collection, TypeInformation)");
 		}
